move social icons bar to layout component
